Create Steganography
This webpage is part of a project where students can have some fun and excitement. It has a title that talks about "Independent Study - Security Homework" and reminds students not to submit assignments late. There's also an encouraging message saying, "YOU GOT THIS!!!" to make students feel confident. The page gives a hint, "Hint: Steganography...", which hints at the exciting topic of the project. What's interesting is that there are hidden things on this page, like links to problems and the first clue. To find them, students need to look at the webpage's code. It's like a little game where you have to search for clues, making learning more interesting and enjoyable.