Create Challenges
This webpage is like the starting point of an exciting puzzle challenge for students. It's called "Independent Study," and it sets the stage for an adventure. Although it may seem simple, there's a hidden task waiting to be discovered. It says "INSTRUCTION ON HOW TO DO THIS IS LOADING......," suggesting that there's a mystery to be unraveled. Below, you'll find three sets of numbers and letters. These might hold the key to the next steps of the puzzle. To fully understand the puzzle and proceed, students will need to inspect the webpage's code. It's like a secret codebreaking game, where you need to explore the code to find the clues and solve the puzzle.