Update Challenge3
This code has two functions for making a secret message not secret. The first one, called caesar_decrypt, needs three things: the secret numbers (called encoded_numbers), how much to move the letters in the secret code (called shift), and how long the alphabet is (called alphabet_length). This function does a few steps. It takes the secret numbers and turns them into real numbers. Then it does the opposite of the secret code by moving the letters back. If the number becomes less than 1, it wraps around to the end of the alphabet. After all that, it changes the numbers back into letters. The second function, decode_with_multiple_shifts, uses the first function to try many ways to make the message not secret. It tries different shifts to see which one makes sense. In the example given, the code uses these functions to figure out what a set of secret numbers means with different shifts. The alphabet for this is "ABCDEFGHIJKLMNOPQRSTUVWXYZ .". The messages for each way of making it not secret are then shown. This helps to understand how to use these functions to find the real message behind a secret code using the Caesar cipher in Python.